[#88488982] Faster deployment using Docker.

+++ docker/recipes/docker-deploy.rb
@@ -1,10 +1,10 @@
 include_recipe 'deploy'
 
 node[:deploy].each do |application, deploy|
-  if node[:opsworks][:instance][:layers].first != deploy[:environment_variables][:layer]
-    Chef::Log.debug("Skipping deploy::docker application #{application} as it is not deployed to this layer")
-    next
-  end
+  # if node[:opsworks][:instance][:layers].first != deploy[:environment_variables][:layer]
+  #   Chef::Log.debug("Skipping deploy::docker application #{application} as it is not deployed to this layer")
+  #   next
+  # end
 
   opsworks_deploy_dir do
     user deploy[:user]
@@ -17,64 +17,66 @@ node[:deploy].each do |application, deploy|
     app application
   end
 
+  Chef::Log.info("docker login")
+  bash "docker login" do
+    user "root"
+    cwd "#{deploy[:deploy_to]}"
+    code <<-EOH
+      docker login -e #{node[:custom_env][:vingle][:DOCKER_EMAIL]} -u #{node[:custom_env][:vingle][:DOCKER_USERNAME]} -p #{node[:custom_env][:vingle][:DOCKER_PASSWORD]}
+    EOH
+  end
+
+  Chef::Log.info("docker cleanup")
   bash "docker-cleanup" do
     user "root"
     code <<-EOH
-      if docker ps | grep #{deploy[:application]}; 
+      if docker ps -a | grep #{deploy[:application]};
       then
         docker stop #{deploy[:application]}
         sleep 3
         docker rm #{deploy[:application]}
         sleep 3
       fi
-      if docker images | grep #{deploy[:application]}; 
+      if docker images | grep #{deploy[:application]};
       then
-        docker rmi #{deploy[:application]}
+        docker rmi #{deploy[:application]}/balmbees:#{node[:custom_env][:vingle][:RAILS_ENV]}
+      fi
+      if docker ps -a | grep td_agent;
+      then
+        :
+      else
+        docker pull #{deploy[:application]}/dockerfiles:td_agent
       fi
     EOH
   end
 
-  # bash "docker-build" do
-  #   user "root"
-  #   cwd "#{deploy[:deploy_to]}/current"
-  #   code <<-EOH
-  #    docker build -t=#{deploy[:application]} . > #{deploy[:application]}-docker.out
-  #   EOH
-  # end
-
-  Chef::Log.info "docker login"
-  Chef::Log.info "docker login -e #{deploy[:environment_variables][:docker_email]} -u #{deploy[:environment_variables][:docker_username]} -p #{deploy[:environment_variables][:docker_password]}"
-  bash "docker login" do
-    user "root"
-    cwd "#{deploy[:deploy_to]}/current"
-    code <<-EOH
-      docker login -e #{deploy[:environment_variables][:docker_email]} -u #{deploy[:environment_variables][:docker_username]} -p #{deploy[:environment_variables][:docker_password]}
-    EOH
-  end
-
-  Chef::Log.info "docker pull images"
-  Chef::Log.info "docker pull #{deploy[:application]}/balmbees"
+  Chef::Log.info("docker pull images")
   bash "docker pull images" do
     user "root"
-    cwd "#{deploy[:deploy_to]}/current"
+    cwd "#{deploy[:deploy_to]}"
     code <<-EOH
-      docker pull #{deploy[:application]}/balmbees
+      docker pull #{deploy[:application]}/balmbees:#{node[:custom_env][:vingle][:RAILS_ENV]}
     EOH
   end
 
   dockerenvs = " "
-  deploy[:environment_variables].each do |key, value|
-    # dockerenvs=dockerenvs+" -e "+key+"="+value
+  node[:custom_env][:vingle].each do |key, value|
     dockerenvs=dockerenvs+" -e \"#{key}=#{value}\""
   end
 
-  Chef::Log.info "docker run"
-  Chef::Log.info "docker run #{dockerenvs} -p #{deploy[:environment_variables][:service_port]}:#{deploy[:environment_variables][:container_port]} -p 8080:8080 --name #{deploy[:application]} -d #{deploy[:application]}/balmbees"
+  Chef::Log.info("docker run #{dockerenvs} --name td -d #{deploy[:application]}/dockerfiles:td_agent")
+  Chef::Log.info("docker run #{dockerenvs} -v /var/log/nginx:/var/log/nginx -p 80:80 -p 8080:8080 --link td:td --name #{deploy[:application]} -d #{deploy[:application]}/balmbees:#{node[:custom_env][:vingle][:RAILS_ENV]}")
   bash "docker-run" do
     user "root"
-    cwd "#{deploy[:deploy_to]}/current"
+    cwd "#{deploy[:deploy_to]}"
     code <<-EOH
-      docker run #{dockerenvs} -p #{deploy[:environment_variables][:service_port]}:#{deploy[:environment_variables][:container_port]} -p 8080:8080 --name #{deploy[:application]} -d #{deploy[:application]}/balmbees
+      if docker ps -a | grep td_agent;
+      then
+        :
+      else
+        docker run #{dockerenvs} --name td -d #{deploy[:application]}/dockerfiles:td_agent
+      fi
+      docker run #{dockerenvs} -v /var/log/nginx:/var/log/nginx -p 80:80 -p 8080:8080 --link td:td --name #{deploy[:application]} -d #{deploy[:application]}/balmbees:#{node[:custom_env][:vingle][:RAILS_ENV]}
     EOH
   end
 end
